Transporting structures from D->R is hard
Seems we also need to do the SEXP stuff if we want to transfer dynamic
structures back from D to R. The pointers are const, and S_realloc
doesn't work for me, also using calloc and overwitting the pointer
has no effect, as its read-only. If I just allocate the expected
return size this works, also our initial idea was to leave all
the data on the D side in a sort of in-memory store, so passing
the arguments like this should be enough.